refactor: macOS theme-related functions and imports - [4/20]
- move the `window_theme_for_background` function to a new file `src/platform/macos/editor.rs`
- remove the `is_light_color` function definition from `src/editor/mod.rs`
- import `window::theme` from `winit` in `src/platform/macos/editor.rs`
