+++ src/platform/macos/mod.rs
@@ -1,3 +1,4 @@
+pub mod editor;
 pub mod keyboard;
 pub mod settings;
